+++ dsp/envelope.hpp
@@ -27,7 +27,7 @@ namespace dspheaders {
         float* times,
         unsigned tLen,
         float samplerate,
-        float (*interpolate)(float, float*, unsigned)
+        float (*interpolate)(float, float*, size_t)
       );
 
       Envelope(
@@ -38,14 +38,14 @@ namespace dspheaders {
         float* curves,
         unsigned cLen,
         float samplerate,
-        float (*interpolate)(float, float*, unsigned)
+        float (*interpolate)(float, float*, size_t)
       );
 
       Envelope(
           float* table,
           unsigned tablelength,
           float samplerate,
-          float (*interpolate)(float, float*, unsigned)
+          float (*interpolate)(float, float*, size_t)
       );
 
       // Returns current value from table
@@ -76,7 +76,7 @@ namespace dspheaders {
     public: 
       float play();
       float play(GATE trigger);
-      PercEnv(float attack, float decay, float samplerate, float (*interpolate)(float, float*, unsigned));
+      PercEnv(float attack, float decay, float samplerate, float (*interpolate)(float, float*, size_t));
   };
 } // namespace dspheaders
   //
