+++ dsp/filter.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include "buffer.hpp"
+#include <cstddef>
 
 namespace dspheaders {
   enum COMBTYPE {
@@ -58,7 +59,7 @@ namespace dspheaders {
       Comb(
           unsigned offset,
           unsigned samplerate,
-          float (*interpolate)(float, float*, unsigned)
+          float (*interpolate)(float, float*, size_t)
       );
   };
 
@@ -80,7 +81,7 @@ namespace dspheaders {
       Allpass(
         unsigned offset,
         unsigned samplerate,
-        float (*interpolate)(float, float*, unsigned)
+        float (*interpolate)(float, float*, size_t)
       );
   };
 }
