+++ dsp/interpolation.hpp
@@ -1,3 +1,5 @@
+#include <cstdlib>
+
 #pragma once
 
 namespace dspheaders {
@@ -9,11 +11,11 @@ namespace dspheaders {
   };
 
   namespace interpolation {
-    float none(float position, float* table, unsigned tablelength);
-    float linear(float position, float* table, unsigned tablelength);
-    float cosine(float position, float* table, unsigned tablelength);
-    float cubic(float position, float* table, unsigned tableLength);
-    float hermetic(float position, float* table, unsigned tableLength);
+    float none(float position, float* table, size_t tablelength);
+    float linear(float position, float* table, size_t tablelength);
+    float cosine(float position, float* table, size_t tablelength);
+    float cubic(float position, float* table, size_t tableLength);
+    float hermetic(float position, float* table, size_t tableLength);
     float slope(float a, float b);
 
   }
